+++ dgraph/cmd/zero/oracle.go
@@ -34,12 +34,15 @@ type syncMark struct {
 
 type Oracle struct {
 	x.SafeMutex
-	commits    map[uint64]uint64 // start -> commit
+	commits map[uint64]uint64 // start -> commit
+	// TODO: Check if we need LRU.
 	rowCommit  map[string]uint64 // fp(key) -> commit
 	aborts     map[uint64]struct{}
 	maxPending uint64
 
-	tmax        uint64
+	tmax uint64
+	// timestamp at the time of start of server or when it became leader.
+	startTxnTs  uint64
 	subscribers map[int]chan *protos.OracleDelta
 	updates     chan *protos.OracleDelta
 	doneUntil   x.WaterMark
@@ -56,7 +59,18 @@ func (o *Oracle) Init() {
 	go o.sendDeltasToSubscribers()
 }
 
+func (o *Oracle) updateStartTxnTs(ts uint64) {
+	o.Lock()
+	defer o.Unlock()
+	o.startTxnTs = ts
+	o.rowCommit = make(map[string]uint64)
+}
+
 func (o *Oracle) hasConflict(src *protos.TxnContext) bool {
+	// This transaction was not started after i became leader.
+	if src.StartTs < o.startTxnTs {
+		return true
+	}
 	for _, k := range src.Keys {
 		if last := o.rowCommit[k]; last > src.StartTs {
 			return true
@@ -66,7 +80,8 @@ func (o *Oracle) hasConflict(src *protos.TxnContext) bool {
 }
 
 func (o *Oracle) purgeBelow(minTs uint64) {
-	x.Printf("purging below %d %d %d\n", minTs, len(o.commits), len(o.aborts))
+	x.Printf("purging below ts:%d, len(o.commits):%d, len(o.aborts):%d\n",
+		minTs, len(o.commits), len(o.aborts))
 	o.Lock()
 	defer o.Unlock()
 
@@ -79,6 +94,13 @@ func (o *Oracle) purgeBelow(minTs uint64) {
 	for ts := range o.aborts {
 		if ts < minTs {
 			delete(o.aborts, ts)
+		}
+	}
+	// There is no transaction running with startTs less than minTs
+	// So we can delete everything from rowCommit whose commitTs < minTs
+	for key, ts := range o.rowCommit {
+		if ts < minTs {
+			delete(o.rowCommit, key)
 		}
 	}
 	o.tmax = minTs
@@ -289,15 +311,18 @@ func (s *Server) CommitOrAbort(ctx context.Context, src *protos.TxnContext) (*pr
 }
 
 var errClosed = errors.New("Streaming closed by Oracle.")
+var errNotLeader = errors.New("Node is no longer leader.")
 
 func (s *Server) Oracle(unused *protos.Payload, server protos.Zero_OracleServer) error {
-	// TODO: Add leader check and test leader changes.
 	ch, id := s.orc.newSubscriber()
 	defer s.orc.removeSubscriber(id)
 
 	ctx := server.Context()
+	leaderChangeCh := s.leaderChangeChannel()
 	for {
 		select {
+		case <-leaderChangeCh:
+			return errNotLeader
 		case delta, open := <-ch:
 			if !open {
 				return errClosed