+++ posting/mvcc.go
@@ -87,6 +87,30 @@ func (t *transactions) MinTs() uint64 {
 	return minTs
 }
 
+// Returns startTs of all pending transactions started upto 10000 raft log
+// entries after last snapshot if the memory consumed by all raft log entries
+// is high.
+func (t *transactions) TxnsSinceSnapshot() []uint64 {
+	lastSnapshotIdx := TxnMarks().DoneUntil()
+	var timestamps []uint64
+	t.Lock()
+	defer t.Unlock()
+	numKeys := 0
+	for _, txn := range t.m {
+		lenDelta, index := txn.lenAndStartIdx()
+		if index-lastSnapshotIdx <= 10000 {
+			timestamps = append(timestamps, txn.StartTs)
+		}
+		numKeys += lenDelta
+	}
+	// Users can do transactions which mutates few edges, numKeys gives us
+	// a good estimation of space consumed by raft log entries
+	if numKeys < 10<<20 { // 500MB considering average size of posting to be 50bytes.
+		return nil
+	}
+	return timestamps
+}
+
 func (t *transactions) Reset() {
 	t.Lock()
 	defer t.Unlock()
@@ -106,6 +130,13 @@ func (t *transactions) Iterate(ok func(key []byte) bool) []uint64 {
 		}
 	}
 	return timestamps
+}
+
+func (t *Txn) lenAndStartIdx() (int, uint64) {
+	t.Lock()
+	defer t.Unlock()
+	x.AssertTrue(len(t.Indices) > 0)
+	return len(t.deltas), t.Indices[0]
 }
 
 func (t *Txn) conflicts(ok func(key []byte) bool) bool {
@@ -143,7 +174,9 @@ func (t *Txn) done() {
 	TxnMarks().DoneMany(t.Indices)
 }
 
-func (t *Txn) Index() uint64 {
+// LastIndex returns the index of last prewrite proposal associated with
+// the transaction.
+func (t *Txn) LastIndex() uint64 {
 	t.Lock()
 	defer t.Unlock()
 	if l := len(t.Indices); l > 0 {