+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project will adhere to [Semantic Versioning](http://semver.org/spec/v2.
 
 * Support for removing dead node from quorum.
 
+### Changed
+
+* Dgraph tries to abort long running/abandoned transactions.
+
 ## [0.9.1] - 2017-11-15
 
 ### Changed